* commit 'c87b8839041abca74b6cd295bd8907ab0346f7dd':
  Make valgrind work for arm64 platform

+++ main/coregrind/m_syswrap/syswrap-linux.c
@@ -4372,7 +4372,7 @@ PRE(sys_openat)
       filename is relative to cwd.  */
    if (ML_(safe_to_deref)( (void*)ARG2, 1 )
        && *(Char *)ARG2 != '/'
-       && ARG1 != VKI_AT_FDCWD
+       && ((Int)ARG1) != VKI_AT_FDCWD
        && !ML_(fd_allowed)(ARG1, "openat", tid, False))
       SET_STATUS_Failure( VKI_EBADF );
